+++ package/CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2018-10-08
+## [0.1.0-preview.2] - 2019.02.22
+
+### Rebuild meta files.
+
+## [0.1.0-preview.1] - 2019-02-01
+
+### Move samples outside of main package.
+
+## [0.1.0-preview.0] - 2018-10-08
 
 ### This is the first release of *Unity Package Settings Manager*.